Rework device-sharing mechanism to volcano
