+++ pkg/scheduler/api/node_info.go
@@ -19,13 +19,13 @@ package api
 import (
 	"fmt"
 	"strconv"
-	"strings"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/klog"
 	k8sframework "k8s.io/kubernetes/pkg/scheduler/framework"
 
 	"volcano.sh/apis/pkg/apis/scheduling/v1beta1"
+	"volcano.sh/volcano/pkg/scheduler/api/devices/nvidia/gpushare"
 )
 
 type AllocateFailError struct {
@@ -70,8 +70,8 @@ type NodeInfo struct {
 	RevocableZone     string
 
 	// Used to store custom information
-	Others     map[string]interface{}
-	GPUDevices map[int]*GPUDevice
+	Others        map[string]interface{}
+	SharedDevices map[string]SharedDevicePool
 
 	// enable node resource oversubscription
 	OversubscriptionNode bool
@@ -140,8 +140,8 @@ func NewNodeInfo(node *v1.Node) *NodeInfo {
 		OversubscriptionResource: EmptyResource(),
 		Tasks:                    make(map[TaskID]*TaskInfo),
 
-		GPUDevices:  make(map[int]*GPUDevice),
-		ImageStates: make(map[string]*k8sframework.ImageStateSummary),
+		SharedDevices: make(map[string]SharedDevicePool),
+		ImageStates:   make(map[string]*k8sframework.ImageStateSummary),
 	}
 
 	nodeInfo.setOversubscription(node)
@@ -153,7 +153,7 @@ func NewNodeInfo(node *v1.Node) *NodeInfo {
 		nodeInfo.Allocatable = NewResource(node.Status.Allocatable).Add(nodeInfo.OversubscriptionResource)
 		nodeInfo.Capability = NewResource(node.Status.Capacity).Add(nodeInfo.OversubscriptionResource)
 	}
-	nodeInfo.setNodeGPUInfo(node)
+	nodeInfo.SharedDevices[GPUSharingDevice] = gpushare.NewGPUDevices(nodeInfo.Name, node)
 	nodeInfo.setNodeState(node)
 	nodeInfo.setRevocableZone(node)
 
@@ -317,38 +317,6 @@ func (ni *NodeInfo) setNodeState(node *v1.Node) {
 	klog.V(4).Infof("set the node %s status to %s.", node.Name, Ready.String())
 }
 
-func (ni *NodeInfo) setNodeGPUInfo(node *v1.Node) {
-	if node == nil {
-		return
-	}
-	memory, ok := node.Status.Capacity[VolcanoGPUResource]
-	if !ok {
-		return
-	}
-	totalMemory := memory.Value()
-
-	res, ok := node.Status.Capacity[VolcanoGPUNumber]
-	if !ok {
-		return
-	}
-	gpuNumber := res.Value()
-	if gpuNumber == 0 {
-		klog.Warningf("invalid %s=%s", VolcanoGPUNumber, res.String())
-		return
-	}
-
-	memoryPerCard := uint(totalMemory / gpuNumber)
-	ni.GPUDevices = make(map[int]*GPUDevice)
-	for i := 0; i < int(gpuNumber); i++ {
-		ni.GPUDevices[i] = NewGPUDevice(i, memoryPerCard)
-	}
-	unhealthyGPUs := ni.getUnhealthyGPUs(node)
-	for i := range unhealthyGPUs {
-		klog.V(4).Infof("delete unhealthy gpu id %d from GPUDevices", unhealthyGPUs[i])
-		delete(ni.GPUDevices, unhealthyGPUs[i])
-	}
-}
-
 // SetNode sets kubernetes node object to nodeInfo object
 func (ni *NodeInfo) SetNode(node *v1.Node) {
 	ni.setNodeState(node)
@@ -376,7 +344,7 @@ func (ni *NodeInfo) SetNode(node *v1.Node) {
 // setNode sets kubernetes node object to nodeInfo object without assertion
 func (ni *NodeInfo) setNode(node *v1.Node) {
 	ni.setOversubscription(node)
-	ni.setNodeGPUInfo(node)
+	ni.SharedDevices[GPUSharingDevice] = gpushare.NewGPUDevices(ni.Name, node)
 	ni.setRevocableZone(node)
 
 	ni.Name = node.Name
@@ -395,13 +363,13 @@ func (ni *NodeInfo) setNode(node *v1.Node) {
 			ni.Idle.sub(ti.Resreq) // sub without assertion
 			ni.Releasing.Add(ti.Resreq)
 			ni.Used.Add(ti.Resreq)
-			ni.AddGPUResource(ti.Pod)
+			ni.SharedDevices[GPUSharingDevice].AddResource(ti.Pod)
 		case Pipelined:
 			ni.Pipelined.Add(ti.Resreq)
 		default:
 			ni.Idle.sub(ti.Resreq) // sub without assertion
 			ni.Used.Add(ti.Resreq)
-			ni.AddGPUResource(ti.Pod)
+			ni.SharedDevices[GPUSharingDevice].AddResource(ti.Pod)
 		}
 	}
 }
@@ -445,7 +413,7 @@ func (ni *NodeInfo) AddTask(task *TaskInfo) error {
 			}
 			ni.Releasing.Add(ti.Resreq)
 			ni.Used.Add(ti.Resreq)
-			ni.AddGPUResource(ti.Pod)
+			ni.SharedDevices[GPUSharingDevice].AddResource(ti.Pod)
 		case Pipelined:
 			ni.Pipelined.Add(ti.Resreq)
 		default:
@@ -453,7 +421,7 @@ func (ni *NodeInfo) AddTask(task *TaskInfo) error {
 				return err
 			}
 			ni.Used.Add(ti.Resreq)
-			ni.AddGPUResource(ti.Pod)
+			ni.SharedDevices[GPUSharingDevice].AddResource(ti.Pod)
 		}
 	}
 
@@ -488,13 +456,13 @@ func (ni *NodeInfo) RemoveTask(ti *TaskInfo) error {
 			ni.Releasing.Sub(task.Resreq)
 			ni.Idle.Add(task.Resreq)
 			ni.Used.Sub(task.Resreq)
-			ni.SubGPUResource(ti.Pod)
+			ni.SharedDevices[GPUSharingDevice].SubResource(ti.Pod)
 		case Pipelined:
 			ni.Pipelined.Sub(task.Resreq)
 		default:
 			ni.Idle.Add(task.Resreq)
 			ni.Used.Sub(task.Resreq)
-			ni.SubGPUResource(ti.Pod)
+			ni.SharedDevices[GPUSharingDevice].SubResource(ti.Pod)
 		}
 	}
 
@@ -545,95 +513,6 @@ func (ni *NodeInfo) Pods() (pods []*v1.Pod) {
 		pods = append(pods, t.Pod)
 	}
 
-	return
-}
-
-// GetDevicesIdleGPUMemory returns all the idle GPU memory by gpu card.
-func (ni *NodeInfo) GetDevicesIdleGPUMemory() map[int]uint {
-	devicesAllGPUMemory := ni.getDevicesAllGPUMemory()
-	devicesUsedGPUMemory := ni.getDevicesUsedGPUMemory()
-	res := map[int]uint{}
-	for id, allMemory := range devicesAllGPUMemory {
-		if usedMemory, found := devicesUsedGPUMemory[id]; found {
-			res[id] = allMemory - usedMemory
-		} else {
-			res[id] = allMemory
-		}
-	}
-	return res
-}
-
-func (ni *NodeInfo) getDevicesUsedGPUMemory() map[int]uint {
-	res := map[int]uint{}
-	for _, device := range ni.GPUDevices {
-		res[device.ID] = device.getUsedGPUMemory()
-	}
-	return res
-}
-
-func (ni *NodeInfo) getDevicesAllGPUMemory() map[int]uint {
-	res := map[int]uint{}
-	for _, device := range ni.GPUDevices {
-		res[device.ID] = device.Memory
-	}
-	return res
-}
-
-// GetDevicesIdleGPU returns all the idle gpu card.
-func (ni *NodeInfo) GetDevicesIdleGPUs() []int {
-	res := []int{}
-	for _, device := range ni.GPUDevices {
-		if device.isIdleGPU() {
-			res = append(res, device.ID)
-		}
-	}
-	return res
-}
-
-// AddGPUResource adds the pod to GPU pool if it is assigned
-func (ni *NodeInfo) AddGPUResource(pod *v1.Pod) {
-	gpuRes := GetGPUMemoryOfPod(pod)
-	if gpuRes > 0 {
-		ids := GetGPUIndex(pod)
-		for _, id := range ids {
-			if dev := ni.GPUDevices[id]; dev != nil {
-				dev.PodMap[string(pod.UID)] = pod
-			}
-		}
-	}
-}
-
-// SubGPUResource frees the gpu hold by the pod
-func (ni *NodeInfo) SubGPUResource(pod *v1.Pod) {
-	gpuRes := GetGPUMemoryOfPod(pod)
-	if gpuRes > 0 {
-		ids := GetGPUIndex(pod)
-		for _, id := range ids {
-			if dev := ni.GPUDevices[id]; dev != nil {
-				delete(dev.PodMap, string(pod.UID))
-			}
-		}
-	}
-}
-
-// getUnhealthyGPUs returns all the unhealthy GPU id.
-func (ni *NodeInfo) getUnhealthyGPUs(node *v1.Node) (unhealthyGPUs []int) {
-	unhealthyGPUs = []int{}
-	devicesStr, ok := node.Annotations[UnhealthyGPUIDs]
-
-	if !ok {
-		return
-	}
-
-	idsStr := strings.Split(devicesStr, ",")
-	for _, sid := range idsStr {
-		id, err := strconv.Atoi(sid)
-		if err != nil {
-			klog.Warningf("Failed to parse unhealthy gpu id %s due to %v", sid, err)
-		} else {
-			unhealthyGPUs = append(unhealthyGPUs, id)
-		}
-	}
 	return
 }
