+++ pkg/scheduler/api/node_info_test.go
@@ -23,6 +23,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8sframework "k8s.io/kubernetes/pkg/scheduler/framework"
+	"volcano.sh/volcano/pkg/scheduler/api/devices/nvidia/gpushare"
 )
 
 func nodeInfoEqual(l, r *NodeInfo) bool {
@@ -65,7 +66,9 @@ func TestNodeInfo_AddPod(t *testing.T) {
 					"c1/p1": NewTaskInfo(case01Pod1),
 					"c1/p2": NewTaskInfo(case01Pod2),
 				},
-				GPUDevices:  make(map[int]*GPUDevice),
+				SharedDevices: map[string]SharedDevicePool{
+					GPUSharingDevice: gpushare.NewGPUDevices("n1", case01Node),
+				},
 				ImageStates: make(map[string]*k8sframework.ImageStateSummary),
 			},
 		},
@@ -86,8 +89,10 @@ func TestNodeInfo_AddPod(t *testing.T) {
 				ResourceUsage:            &NodeUsage{},
 				State:                    NodeState{Phase: Ready},
 				Tasks:                    map[TaskID]*TaskInfo{},
-				GPUDevices:               make(map[int]*GPUDevice),
-				ImageStates:              make(map[string]*k8sframework.ImageStateSummary),
+				SharedDevices: map[string]SharedDevicePool{
+					GPUSharingDevice: gpushare.NewGPUDevices("n2", case01Node),
+				},
+				ImageStates: make(map[string]*k8sframework.ImageStateSummary),
 			},
 			expectedFailure: true,
 		},
@@ -149,7 +154,9 @@ func TestNodeInfo_RemovePod(t *testing.T) {
 					"c1/p1": NewTaskInfo(case01Pod1),
 					"c1/p3": NewTaskInfo(case01Pod3),
 				},
-				GPUDevices:  make(map[int]*GPUDevice),
+				SharedDevices: map[string]SharedDevicePool{
+					GPUSharingDevice: gpushare.NewGPUDevices("n1", case01Node),
+				},
 				ImageStates: make(map[string]*k8sframework.ImageStateSummary),
 			},
 		},
@@ -212,7 +219,9 @@ func TestNodeInfo_SetNode(t *testing.T) {
 					"c1/p2": NewTaskInfo(case01Pod2),
 					"c1/p3": NewTaskInfo(case01Pod3),
 				},
-				GPUDevices:  make(map[int]*GPUDevice),
+				SharedDevices: map[string]SharedDevicePool{
+					GPUSharingDevice: gpushare.NewGPUDevices("n1", case01Node1),
+				},
 				ImageStates: make(map[string]*k8sframework.ImageStateSummary),
 			},
 		},
